Markdown形式のドキュメント生成機能を追加。
