+++ DeviceConnectCodegen/modules/deviceconnect-codegen/src/main/java/org/deviceconnect/codegen/DConnectCodegen.java
@@ -13,6 +13,7 @@ import io.swagger.parser.SwaggerParser;
 import org.apache.commons.cli.*;
 import org.deviceconnect.codegen.app.HtmlAppCodegenConfig;
 import org.deviceconnect.codegen.docs.HtmlDocsCodegenConfig;
+import org.deviceconnect.codegen.docs.MarkdownDocsCodegenConfig;
 import org.deviceconnect.codegen.plugin.AndroidPluginCodegenConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -248,7 +249,8 @@ public class DConnectCodegen {
     }
 
     private static void checkProfileName(final DConnectCodegenConfig config, final String profileName) {
-        if (config instanceof HtmlAppCodegenConfig || config instanceof HtmlDocsCodegenConfig) {
+        if (config instanceof HtmlAppCodegenConfig || config instanceof HtmlDocsCodegenConfig ||
+            config instanceof MarkdownDocsCodegenConfig) {
             return;
         }
