+++ DeviceConnectCodegen/modules/deviceconnect-codegen/src/main/java/org/deviceconnect/codegen/ProfileTemplate.java
@@ -1,0 +1,8 @@
+package org.deviceconnect.codegen;
+
+
+public class ProfileTemplate {
+
+    public String templateFile;
+    public String outputFile;
+}
