+++ DeviceConnectCodegen/modules/deviceconnect-codegen/src/main/java/org/deviceconnect/codegen/plugin/IosPluginCodegenConfig.java
@@ -8,6 +8,7 @@ import io.swagger.models.parameters.FormParameter;
 import io.swagger.models.parameters.Parameter;
 import io.swagger.models.parameters.QueryParameter;
 import io.swagger.models.properties.*;
+import org.deviceconnect.codegen.ProfileTemplate;
 
 import java.io.File;
 import java.util.ArrayList;
