--- DeviceConnectCodegen/modules/deviceconnect-codegen/src/main/java/org/deviceconnect/codegen/plugin/ProfileTemplate.java
@@ -1,9 +1,0 @@
-package org.deviceconnect.codegen.plugin;
-
-
-class ProfileTemplate {
-
-    String templateFile;
-    String outputFile;
-
-}
