+++ DeviceConnectCodegen/modules/deviceconnect-codegen/src/main/resources/META-INF/services/org.deviceconnect.codegen.DConnectCodegenConfig
@@ -1,5 +1,6 @@
 org.deviceconnect.codegen.app.HtmlAppCodegenConfig
 org.deviceconnect.codegen.docs.HtmlDocsCodegenConfig
+org.deviceconnect.codegen.docs.MarkdownDocsCodegenConfig
 org.deviceconnect.codegen.plugin.AndroidPluginCodegenConfig
 org.deviceconnect.codegen.plugin.IosPluginCodegenConfig
 org.deviceconnect.codegen.plugin.NodePluginCodegenConfig
