+++ DeviceConnectCodegen/modules/deviceconnect-codegen/src/main/resources/deviceConnectMarkdownDocs/dconnect-message.mustache
@@ -1,0 +1,8 @@
+<table>
+    <tbody>
+        <tr><th colspan="{{maxNestLevel}}">物理名</th><th>データ型</th><th>必須</th><th>説明</th></tr>
+        {{#paramList}}
+        <tr>{{#indents}}<td></td>{{/indents}}<td colspan="{{colSpan}}">{{name}}</td><td>{{dataType}}</td><td>{{required}}</td><td>{{#title}}<b>{{{.}}}</b><br>{{/title}}{{{description}}}</td></tr>
+        {{/paramList}}
+    </tbody>
+</table>
