+++ DeviceConnectCodegen/samples/md-docs.sh
@@ -1,0 +1,20 @@
+#!/bin/sh -x
+
+# スケルトンコード種別: APIリファレンス (Markdown形式)
+LANG="deviceConnectMarkdownDocs"
+
+# プロファイル定義ファイル
+SPEC="../standard-profile-specs"
+
+# スケルトンコード出力先
+OUTPUT_DIR="./output/markdown/Device_Connect_RESTful_API_Specification"
+
+# API仕様書の表示名
+DISPLAY_NAME="Device_Connect_RESTful_API_Specification"
+
+# スケルトンコード生成ツールのバイナリ
+JAR_FILE="../bin/deviceconnect-codegen.jar"
+
+ARGS="--input-spec-dir $SPEC  --lang $LANG  --display-name $DISPLAY_NAME  --output $OUTPUT_DIR"
+
+java -Dfile.encoding=UTF-8 -jar $JAR_FILE $ARGS
